requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.23.1
bottle==0.12.23
bottle-websocket==0.2.9
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Eel==0.14.0
Flask==2.2.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pefile==2022.5.30
pycparser==2.21
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
pywin32-ctypes==0.2.0
speedtest-cli==2.1.3
Werkzeug==2.2.2
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
